Committed Ron Kuris's warning fixes.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1097 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/notationhlayout.cpp
@@ -471,7 +471,7 @@ NotationHLayout::reconcileBars()
 
 	    BarDataList &list = i->second;
 
-	    if ((int)list.size() > barNo) {
+	    if (list.size() > barNo) {
 
 		reachedEnd = false;
 
@@ -485,7 +485,7 @@ NotationHLayout::reconcileBars()
 
 	    BarDataList &list = i->second;
 
-	    if ((int)list.size() > barNo) {
+	    if (list.size() > barNo) {
 
 		BarData &bd(list[barNo]);
